+++ src/modules/goals/goals.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { DatabaseModule } from 'src/database/database.module';
+import { GoalsController } from './goals.controller';
+import GoalsService from './goals.service';
+
+@Module({
+
+  imports: [DatabaseModule],
+  providers: [GoalsService],
+  controllers: [GoalsController]
+})
+export class GoalsModule { }